Walls Work

+++ core/src/com/snake/game/Grid.java
@@ -50,9 +50,7 @@ public class Grid {
             int posX = random.nextInt(gridSize.x);
             int posY = random.nextInt(gridSize.y);
 
-            
-            walls[i] = new Wall(new Vector(posX, posY), null, new Vector(posX * squareSize - (gridSize.x/2) * squareSize, posY * squareSize - (gridSize.y/2) * squareSize), new Vector(random.nextInt(1, 5), random.nextInt(1, 5)));
-
+            walls[i] = new Wall(new Vector(posX, posY), null, new Vector((int)(posX * squareSize - (gridSize.x/2.0) * squareSize), (int)(posY * squareSize - (gridSize.y/2.0) * squareSize)), new Vector(random.nextInt(1, 5), random.nextInt(1, 5)));
 
             if(walls[i].size.x + posX > gridSize.x){
                 walls[i].size.x = gridSize.x - posX;
